Starting the project architecture refactoring
