+++ seapopym/configuration/no_transport/configuration_to_dataset.py
@@ -8,7 +8,7 @@ import attrs
 import numpy as np
 import xarray as xr
 
-from seapopym.configuration.no_transport.parameter_functional_group import (
+from seapopym.configuration.parameters.parameter_functional_group import (
     FunctionalGroupUnit,
     FunctionalGroupUnitMigratoryParameters,
     FunctionalGroupUnitRelationParameters,
