+++ seapopym/configuration/parameters/parameter_functional_group.py
