--- seapopym/post_production/no_transport/biomass.py
@@ -1,1 +1,0 @@
-"""Wrap the biomass computation to follow the no-transport model."""
