--- seapopym/production/no_transport/production.py
@@ -1,1 +1,0 @@
-"""Wrap the biomass computation to follow the no-transport model."""
