+++ test/configuration/no_transport/test_configuration.py
@@ -2,15 +2,15 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from seapopym.configuration.no_transport import parameter_functional_group
+from seapopym.configuration.parameters import parameter_functional_group
 from seapopym.configuration.no_transport.configuration import NoTransportConfiguration
 from seapopym.configuration.no_transport.parameter import (
     ForcingParameters,
     FunctionalGroups,
     NoTransportParameters,
 )
-from seapopym.configuration.no_transport.parameter_environment import EnvironmentParameter
-from seapopym.configuration.no_transport.parameter_forcing import ForcingUnit
+from seapopym.configuration.parameters.parameter_environment import EnvironmentParameter
+from seapopym.configuration.parameters.parameter_forcing import ForcingUnit
 from seapopym.standard import coordinates
 from seapopym.standard.units import StandardUnitsLabels
